Multiple fixes

+++ include/ExVectrMath/matrix_types.h
@@ -14,13 +14,13 @@ namespace Math {
 /**
  * Creates a matrix with given size and floats as internal type.
  */
-template<uint16_t ROWS, uint16_t COLS>
+template<size_t ROWS, size_t COLS>
 using Matrix_F = Matrix<float, ROWS, COLS>;
 
 /**
  * Creates a matrix with given size and doubles as internal type.
  */
-template<uint16_t ROWS, uint16_t COLS>
+template<size_t ROWS, size_t COLS>
 using Matrix_D = Matrix<double, ROWS, COLS>;
